Introduce psql-specific datastore implementation
Datastore is able to handle data transfer from/to the specified
database. The database and its configuration is dictated by the
corresponding build-time tags.
Using psql for example, the PostgreSQL implementation is used in
the application.
This will also affect the primary key of the POI model, since the
primary key type is aliased in order to support multiple database
providers.